refactor(api-proxy): split proxy-request into focused modules
