+++ containers/api-proxy/proxy-request.js
@@ -14,6 +14,14 @@ const metrics = require('./metrics');
 const rateLimiter = require('./rate-limiter');
 const { buildUpstreamPath, shouldStripHeader } = require('./proxy-utils');
 const { sanitizeNullToolCallTypes, injectSteeringMessage, injectStreamOptions } = require('./body-transform');
+const {
+  maybeStripLearnedHeaderValues,
+  resetDeprecatedHeaderValuesForTests,
+  parseDeprecatedHeaderFromBody,
+  learnAndStripDeprecatedHeaderValue,
+} = require('./deprecated-header-tracker');
+const { extractBillingHeaders } = require('./billing-headers');
+const { createUpstreamResponseHandlers } = require('./upstream-response');
 const { createRateLimitChecker } = require('./rate-limit');
 const { createProxyWebSocket } = require('./websocket-proxy');
 const {
@@ -85,103 +93,6 @@ const limiter = rateLimiter.create();
 /** When false, token-budget warnings are never injected into request bodies. */
 const isSteeringEnabled = () => process.env.AWF_ENABLE_TOKEN_STEERING === 'true';
 
-// ── Deprecated header value handling ──────────────────────────────────────────
-// General mechanism: when an upstream returns a 400 indicating that a specific
-// value in a specific header is not accepted, we learn that (header, value) pair,
-// strip it, retry once, and proactively remove it from all subsequent requests.
-
-/** Map of headerName → Set of rejected values, learned from upstream 400 responses. */
-const deprecatedHeaderValues = new Map();
-const MAX_CACHED_VALUES_PER_HEADER = 200;
-
-/**
- * Pattern to detect header-value rejection errors from Anthropic.
- * Matches: Unexpected value(s) `<value>` for the `<header>` header
- */
-const DEPRECATED_HEADER_PATTERN = /Unexpected value\(s\)\s+`([^`]+)`\s+for the `([^`]+)` header/;
-
-function normalizeHeaderValue(value) {
-  if (!value) return '';
-  return Array.isArray(value) ? value.join(',') : String(value);
-}
-
-function splitHeaderValue(value) {
-  return normalizeHeaderValue(value).split(',').map(s => s.trim()).filter(Boolean);
-}
-
-function updateHeader(headers, headerName, values) {
-  if (!values.length) {
-    delete headers[headerName];
-    return;
-  }
-  headers[headerName] = values.join(',');
-}
-
-function stripValuesFromHeader(headers, headerName, valuesToStrip) {
-  if (!headers[headerName] || !valuesToStrip.size) return null;
-  const existingValues = splitHeaderValue(headers[headerName]);
-  if (!existingValues.length) {
-    delete headers[headerName];
-    return { removed: [], remaining: [] };
-  }
-  const remaining = existingValues.filter(value => !valuesToStrip.has(value));
-  const removed = existingValues.filter(value => valuesToStrip.has(value));
-  if (!removed.length) return null;
-  updateHeader(headers, headerName, remaining);
-  return { removed, remaining };
-}
-
-function getDeprecatedValuesForHeader(headerName) {
-  if (!deprecatedHeaderValues.has(headerName)) {
-    deprecatedHeaderValues.set(headerName, new Set());
-  }
-  return deprecatedHeaderValues.get(headerName);
-}
-
-function maybeStripLearnedHeaderValues(headers, requestId, provider) {
-  for (const [headerName, rejectedValues] of deprecatedHeaderValues) {
-    if (!headers[headerName] || !rejectedValues.size) continue;
-    const stripped = stripValuesFromHeader(headers, headerName, rejectedValues);
-    if (!stripped) continue;
-    logRequest('warn', 'deprecated_header_stripped', {
-      request_id: requestId,
-      provider,
-      header: headerName,
-      mode: 'cached',
-      removed_values: stripped.removed,
-      remaining_values: stripped.remaining,
-      message: `Removed deprecated ${headerName} values learned from prior upstream 400 responses`,
-    });
-  }
-}
-
-function parseDeprecatedHeaderFromBody(body) {
-  const match = body.toString('utf8').match(DEPRECATED_HEADER_PATTERN);
-  if (!match) return null;
-  return { value: match[1].trim(), header: match[2].trim() };
-}
-
-function learnAndStripDeprecatedHeaderValue(headers, headerName, deprecatedValue, requestId, provider) {
-  const rejectedValues = getDeprecatedValuesForHeader(headerName);
-  rejectedValues.add(deprecatedValue);
-  if (rejectedValues.size > MAX_CACHED_VALUES_PER_HEADER) {
-    const oldest = rejectedValues.values().next().value;
-    if (oldest !== undefined) rejectedValues.delete(oldest);
-  }
-  const stripped = stripValuesFromHeader(headers, headerName, new Set([deprecatedValue]));
-  if (!stripped) return null;
-  logRequest('warn', 'deprecated_header_stripped', {
-    request_id: requestId,
-    provider,
-    header: headerName,
-    mode: 'retry',
-    removed_values: stripped.removed,
-    remaining_values: stripped.remaining,
-    message: `Removed deprecated ${headerName} value rejected by upstream: ${deprecatedValue}`,
-  });
-  return stripped;
-}
-
 function getUrlPathForSpan(requestUrl) {
   if (typeof requestUrl !== 'string' || !requestUrl) return '/';
   try {
@@ -189,50 +100,6 @@ function getUrlPathForSpan(requestUrl) {
   } catch {
     return '/';
   }
-}
-
-// ── Billing header extraction ─────────────────────────────────────────────────
-
-/**
- * Extract billing/quota information from upstream response headers.
- *
- * CAPI returns quota snapshots as `X-Quota-Snapshot-<Type>` headers with
- * URL-encoded fields: ent (entitlement), ov (overage), ovPerm (overage allowed),
- * rem (remaining %), rst (reset date).
- *
- * Also captures X-RateLimit-* headers from CAPI responses.
- *
- * @param {Record<string, string|string[]>} headers - Response headers
- * @returns {object|null} Billing info object, or null if no billing headers present
- */
-function extractBillingHeaders(headers) {
-  const billing = {};
-  let hasBilling = false;
-
-  for (const [name, value] of Object.entries(headers)) {
-    const lower = name.toLowerCase();
-    if (lower.startsWith('x-quota-snapshot-')) {
-      const quotaType = lower.slice('x-quota-snapshot-'.length);
-      try {
-        const params = new URLSearchParams(String(value));
-        const snapshot = {};
-        for (const [k, v] of params) snapshot[k] = v;
-        billing[`quota_${quotaType}`] = snapshot;
-      } catch {
-        billing[`quota_${quotaType}_raw`] = String(value);
-      }
-      hasBilling = true;
-    }
-  }
-
-  if (headers['x-ratelimit-limit']) {
-    billing.rate_limit = headers['x-ratelimit-limit'];
-    billing.rate_remaining = headers['x-ratelimit-remaining'];
-    billing.rate_reset = headers['x-ratelimit-reset'];
-    hasBilling = true;
-  }
-
-  return hasBilling ? billing : null;
 }
 
 // ── Utility ───────────────────────────────────────────────────────────────────
@@ -357,147 +224,19 @@ function buildRequestHeaders(body, inboundBytes, req, { injectHeaders, provider,
   return headers;
 }
 
-/**
- * Log request completion: emit metrics and a structured `request_complete` log.
- *
- * @param {number} statusCode
- * @param {number} responseBytes
- * @param {string|null} initiatorSent
- * @param {object|null} billingInfo
- * @param {{ startTime: number, provider: string, req: object, requestBytes: number, targetHost: string, requestId: string }} ctx
- */
-function logRequestCompletion(statusCode, responseBytes, initiatorSent, billingInfo, {
-  startTime, provider, req, requestBytes, targetHost, requestId,
-}) {
-  const duration = Date.now() - startTime;
-  const sc = metrics.statusClass(statusCode);
-  metrics.gaugeDec('active_requests', { provider });
-  metrics.increment('requests_total', { provider, method: req.method, status_class: sc });
-  metrics.increment('response_bytes_total', { provider }, responseBytes);
-  metrics.observe('request_duration_ms', duration, { provider });
-  if (statusCode >= 200 && statusCode < 300) {
-    applyMaxRunsInvocation();
-  }
-  const logFields = {
-    request_id: requestId, provider, method: req.method,
-    path: sanitizeForLog(req.url), status: statusCode,
-    duration_ms: duration, request_bytes: requestBytes,
-    response_bytes: responseBytes, upstream_host: targetHost,
-  };
-  if (initiatorSent) logFields.x_initiator = initiatorSent;
-  if (billingInfo) logFields.billing = billingInfo;
-  logRequest('info', 'request_complete', logFields);
-}
-
-/**
- * Emit a warning log when the upstream returns an auth-related 4xx status.
- *
- * @param {number} statusCode
- * @param {{ requestId: string, provider: string, targetHost: string, req: object }} ctx
- */
-function logUpstreamAuthError(statusCode, { requestId, provider, targetHost, req }) {
-  if (statusCode === 400 || statusCode === 401 || statusCode === 403) {
-    logRequest('warn', 'upstream_auth_error', {
-      request_id: requestId, provider, status: statusCode,
-      upstream_host: targetHost, path: sanitizeForLog(req.url),
-      message: `Upstream returned ${statusCode} — check that the API key is valid and correctly formatted`,
-    });
-  }
-}
-
-/**
- * Handle the upstream response: stream or buffer it, track tokens, handle errors,
- * and optionally retry once when a deprecated header value caused a 400.
- *
- * @param {import('http').IncomingMessage} proxyRes
- * @param {object} requestHeaders - Headers that were sent with the upstream request
- * @param {{ res: object, provider: string, requestId: string, req: object,
- *           targetHost: string, startTime: number, span: object, requestBytes: number,
- *           hasRetried: boolean, onRetry: Function }} ctx
- */
-function handleUpstreamResponse(proxyRes, requestHeaders, {
-  res, provider, requestId, req, targetHost, startTime, span, requestBytes, hasRetried, onRetry,
-}) {
-  let responseBytes = 0;
-  const billingInfo = extractBillingHeaders(proxyRes.headers);
-  const initiatorSent = requestHeaders['x-initiator'] || null;
-  const shouldBuffer400ForHeaderStrip =
-    (provider === 'anthropic' || provider === 'copilot') &&
-    !hasRetried &&
-    proxyRes.statusCode === 400;
-
-  const completionCtx = { startTime, provider, req, requestBytes, targetHost, requestId };
-  const authErrCtx = { requestId, provider, targetHost, req };
-
-  proxyRes.on('error', (err) => {
-    otel.endSpanError(span, err, 502);
-    handleRequestError(err, {
-      res, requestId, provider, req, targetHost, startTime,
-      statusCode: 502, clientMessage: 'Response stream error',
-      onHeadersSent: () => {
-        if (typeof res.destroy === 'function') res.destroy(err);
-      },
-    });
-  });
-
-  if (shouldBuffer400ForHeaderStrip) {
-    const bufferedChunks = [];
-    proxyRes.on('data', (chunk) => {
-      responseBytes += chunk.length;
-      bufferedChunks.push(chunk);
-    });
-    proxyRes.on('end', () => {
-      const responseBody = Buffer.concat(bufferedChunks);
-      const deprecated = parseDeprecatedHeaderFromBody(responseBody);
-      if (deprecated) {
-        const retryHeaders = { ...requestHeaders };
-        const stripped = learnAndStripDeprecatedHeaderValue(
-          retryHeaders, deprecated.header, deprecated.value, requestId, provider,
-        );
-        if (stripped) {
-          onRetry(retryHeaders);
-          return;
-        }
-      }
-
-      logRequestCompletion(proxyRes.statusCode, responseBytes, initiatorSent, billingInfo, completionCtx);
-      logUpstreamAuthError(proxyRes.statusCode, authErrCtx);
-
-      const resHeaders = {
-        ...proxyRes.headers,
-        'x-request-id': requestId,
-        'content-length': String(responseBody.length),
-      };
-      delete resHeaders['transfer-encoding'];
-      res.writeHead(proxyRes.statusCode, resHeaders);
-      res.end(responseBody);
-      otel.endSpan(span, proxyRes.statusCode);
-    });
-    return;
-  }
-
-  proxyRes.on('data', (chunk) => { responseBytes += chunk.length; });
-  proxyRes.on('end', () => {
-    logRequestCompletion(proxyRes.statusCode, responseBytes, initiatorSent, billingInfo, completionCtx);
-  });
-
-  const resHeaders = { ...proxyRes.headers, 'x-request-id': requestId };
-  logUpstreamAuthError(proxyRes.statusCode, authErrCtx);
-  res.writeHead(proxyRes.statusCode, resHeaders);
-  proxyRes.pipe(res);
-
-  const isStreaming = (proxyRes.headers['content-type'] || '').includes('text/event-stream');
-  trackTokenUsage(proxyRes, {
-    requestId, provider, path: sanitizeForLog(req.url), startTime, metrics, billingInfo, initiatorSent,
-    onUsage: (normalizedUsage, model) => {
-      otel.setTokenAttributes(span, { provider, model, normalizedUsage, streaming: isStreaming });
-      applyEffectiveTokenUsage(normalizedUsage, model);
-    },
-    onSpanEnd: (statusCode) => {
-      otel.endSpan(span, statusCode);
-    },
-  });
-}
+const { handleUpstreamResponse } = createUpstreamResponseHandlers({
+  metrics,
+  logRequest,
+  sanitizeForLog,
+  otel,
+  handleRequestError,
+  trackTokenUsage,
+  applyEffectiveTokenUsage,
+  applyMaxRunsInvocation,
+  extractBillingHeaders,
+  parseDeprecatedHeaderFromBody,
+  learnAndStripDeprecatedHeaderValue,
+});
 
 /**
  * Create and dispatch the upstream HTTPS request.
@@ -757,7 +496,7 @@ module.exports = {
   resetEffectiveTokenGuardForTests,
   resetMaxRunsGuardForTests,
   resetTimeoutSteeringForTests,
-  resetAnthropicDeprecatedBetaHeadersForTests: () => deprecatedHeaderValues.clear(),
+  resetAnthropicDeprecatedBetaHeadersForTests: resetDeprecatedHeaderValuesForTests,
   getAndClearPendingSteeringMessage,
   getAndClearPendingTimeoutSteeringMessage,
   injectSteeringMessage,
